Added content to the files that make up the manual. In addition added very terse descriptions to the general input variables. Finally, made a small modification to the order in which the different input variables are presented in the manual. They now appear in alphabetial order.
One more detail that I just noticed is that there is a problem in mkmanual:
Right now it uses the INSTALLED version of parsnip.corvus.config instead of the
one in src/util. This is problematic snce one has to remember to copy the
edited version back to src once it has been updated.